requirements.txt
asgiref==3.5.1
attrs==21.4.0
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-nested-routers==0.93.4
drf-spectacular==0.22.1
inflection==0.5.1
jsonschema==4.5.1
PyJWT==2.4.0
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
channels==3.0.4
celery==5.2.7
redis==4.3.3